moderator student user classes
