+++ src/com/mie/model/Moderator.java
@@ -1,0 +1,9 @@
+package com.mie.model;
+
+public class Moderator extends User{
+	//admin constructor
+		public Moderator(String username, String password) {
+			super(username,password);
+		}
+		
+}
